+++ CHANGELOG.md
@@ -23,6 +23,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Verification
 - Transfers
 
+[1.0.1] - 2025-05-08
+
+### Added
+
+- Automatic token refresh
+
+### Fixed
+
+- Token expiry
+
 <!-- Example -->
 
 <!-- ## [Unreleased] - Release Date